Switch Superblock::CpidIndex::TotalMagnitude() to return integers
This method returned a floating-point number. This switches the return
value to an integer. The upcoming changes to the voting system rely on
the total magnitude to calculate voting weight. The use of the integer
return type avoids floating-point discrepancies and allows the network
to potentially integrate the voting system into the protocol in future
enhancements.

Note that this change truncates the fractional portion of the network-
wide magnitude from the return value. It is not significant to current
voting weight calculations.

+++ src/neuralnet/superblock.cpp
@@ -733,9 +733,9 @@ size_t Superblock::CpidIndex::TotalCount() const
     return size() + m_zero_magnitude_count;
 }
 
-double Superblock::CpidIndex::TotalMagnitude() const
+uint64_t Superblock::CpidIndex::TotalMagnitude() const
 {
-    return static_cast<double>(m_total_magnitude) / NN::Magnitude::SCALE_FACTOR;
+    return m_total_magnitude / NN::Magnitude::SCALE_FACTOR;
 }
 
 double Superblock::CpidIndex::AverageMagnitude() const
@@ -744,7 +744,7 @@ double Superblock::CpidIndex::AverageMagnitude() const
         return 0;
     }
 
-    return TotalMagnitude() / size();
+    return static_cast<double>(TotalMagnitude()) / size();
 }
 
 Magnitude Superblock::CpidIndex::MagnitudeOf(const Cpid& cpid) const
